--- a/keyvi/src/cpp/dictionary/fsa/internal/json_value_store.h
+++ b/keyvi/src/cpp/dictionary/fsa/internal/json_value_store.h
@@ -27,6 +27,8 @@
 
 #include <boost/functional/hash.hpp>
 #include <zlib.h>
+
+#define RAPIDJSON_SSE42
 #include "rapidjson/document.h"
 #include "rapidjson/writer.h"
 #include "rapidjson/stringbuffer.h"