… keyvi/src/cpp/dictionary/fsa/internal/constants.h
… keyvi/src/cpp/keyviinspector/keyviinspector.cpp
… keyvi/tests/cpp/dictionary/dictionary_test.cpp
… keyvi/tests/cpp/dictionary/fsa/automata_test.cpp