@david-cliqz

This implements GetOutGoingTransitions(...), which is better than the brute force TryWalkTransition(...).

Dumping some file is now 1s instead of 4s, with using the SSE support it is even below 1s (0.8s on my machine).

It would be good to refactor the statetraversers with that, so that completions get faster.