@DavidNemeskey 2015-10-06
Thanks for the nice intrinsics example. :)